% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACE_REF.R
\name{MACE_REF}
\alias{MACE_REF}
\title{MACE REF}
\usage{
MACE_REF(Obj, donor_id_col, Samples_to_annotate = NULL)
}
\arguments{
\item{Obj}{A Seurat object containing the input single-cell data with it's corresponding metadata.}

\item{donor_id_col}{A string identifying the metadata column name where the sample identification is stored, this is important because the method is performed sample by sample.}

\item{Samples_to_annotate}{A list of unique sample ids to perform the cancer annotation for, has to be found in the donor_id_col. if NULL (default), the annotation will be performed for all samples.}
}
\value{
A Seurat Object containing the original input in addition to REF_cancer_annot metadata column containing the MACE annotations.
}
\description{
This function performs the malignant cell (MACE) annotation, identifying each cell being normal or tumor based on the Reference-based method scATOMIC.
}
\examples{
\dontrun{
Object <- MACE_REF(Obj=Object, donor_id_col = "patient_id")
}

}
