% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Peng_PDAC_3}
\alias{Peng_PDAC_3}
\title{Data to showcase the usage of the MACE package}
\format{
An object of class \code{Seurat} with 18008 rows and 3313 columns.
}
\source{
https://www.nature.com/articles/s41422-019-0195-y
}
\usage{
Peng_PDAC_3
}
\description{
Contains a downsampled version from the public Peng_et_al PDAC cancer dataset (3 samples: T1, T4 and T5)
}
\examples{
data(Peng_PDAC_3)
}
\keyword{datasets}
